requirements.txt
langgraph==0.2.16
langchain==0.2.16
langchain-openai==0.1.23
langchain-community==0.2.16
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
streamlit==1.28.1
